fix(tests): cast to bigint to prevent integer overflow in RF1 TPC-H refresh
The RF1 bulk INSERT into orders computes o_custkey as:
  1 + ((__NEXT_ORDERKEY__ + i) * 1013 % __SF_CUSTOMERS__)

At SF-10 with large __NEXT_ORDERKEY__ values (>2M), the multiplication
overflows a 32-bit integer. For example:
  (2115042 + 15000) * 1013 = 2,157,752,546 > max(int32) = 2,147,483,647

This causes 'integer out of range' errors during refresh cycles.

Cast __NEXT_ORDERKEY__ to bigint to perform the arithmetic in 64-bit
before applying modulo reduction, preserving the deterministic hash
behaviour while avoiding overflow.

+++ tests/tpch/rf1.sql
@@ -8,7 +8,7 @@ INSERT INTO orders (o_orderkey, o_custkey, o_orderstatus, o_totalprice, o_orderd
                     o_orderpriority, o_clerk, o_shippriority, o_comment)
 SELECT
     __NEXT_ORDERKEY__ + i AS o_orderkey,
-    1 + ((__NEXT_ORDERKEY__ + i) * 1013 % __SF_CUSTOMERS__) AS o_custkey,
+    1 + ((__NEXT_ORDERKEY__::bigint + i) * 1013 % __SF_CUSTOMERS__) AS o_custkey,
     (ARRAY['O','F','P'])[1 + i % 3] AS o_orderstatus,
     round((10000 + (i * 7919) % 500000)::numeric / 100, 2) AS o_totalprice,
     DATE '1995-01-01' + (i % 1000)::int AS o_orderdate,
